requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
bitsandbytes==0.45.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
-e git+https://github.com/huggingface/diffusers.git@9f28f1abbaf1de21454644ea5391389dabe9a14a#egg=diffusers
docker-pycreds==0.4.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
peft==0.7.0
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
safetensors==0.5.2
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tokenizers==0.21.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.6
wcwidth==0.2.13
Werkzeug==3.1.3
xformers==0.0.29.post2
zipp==3.21.0
